PR changes implemented
Increaded verbosity/readablility with
    m -> responseMap
    handleErr -> if err != nil {...

Fixed indentation in README.md examples

Broke the single example into 2 sepearte examples
